feat: print read file contents
